Adding link to "State of Componentised Web" post
A blog post that covers "components" for a number of high profile JS frameworks and libraries: AngularJS, Ember and React. The point of this post is to demonstrate that building apps out of components is beneficial and that Web Components represent a standardised approach.

Hopefully this fits well into the reading section.